refactor: bug 'cast one of this addition operation to double'

+++ kata/6-kyu/simple-square-numbers/main/Solution.java
@@ -2,6 +2,6 @@ import static java.util.stream.LongStream.range;
 
 interface Solution {
   static long solve(int n) {
-    return range(1, n).map(i -> i * i).filter(i -> Math.sqrt(i + n) % 1 == 0).findFirst().orElse(-1);
+    return range(1, n).map(i -> i * i).filter(i -> Math.sqrt(i + (double) n) % 1 == 0).findFirst().orElse(-1);
   }
 }
